fix render bugs
