+++ client/pages/notfound.js
@@ -1,5 +1,4 @@
 const html = require('choo/html')
-const css = require('sheetify')
 const datIcons = require('dat-icons')
 const header = require('../elements/header')
 const footer = require('../elements/footer')
@@ -7,32 +6,19 @@ const footer = require('../elements/footer')
 module.exports = function (state, emit) {
   emit(state.events.DOMTITLECHANGE, '404 - Dat Not Found')
 
-  const splash = css`
-    :host {
-      background-repeat: repeat-y;
-      background-position: center 0px;
-
-      @media screen and (min-width: 30em) {
-        /* ns - not small breakpoint from tachyons */
-        background-position: center -125px;
-      }
-    }
-  `
-  const backgroundImageUrl = '/public/img/bg-landing-page.svg'
   return html`
     <body>
       ${header()}
     <div>
-      <section class="${splash} bg-neutral pb4 white cover" style="background-image: url(${backgroundImageUrl})">
+      <section class="bg-neutral pb4 white cover">
         <header class="tc ph5 lh-copy">
           <h1 class="f1 f-headline-l code mb3 fw9 dib tracked-tight green">
-            Dat
-            <div class="dib-ns w3 h3 dn center"><svg><use xlink:href="#daticon-sad-dat"/></svg></div>
-            404
+            <span class="dib">Dat</span><div class="ph4 dib-ns dn"><svg class="w3 h3"><use xlink:href="#daticon-sad-dat"/></svg></div>
+            <span class="dib">404</span>
           </h1>
           <h2 class="tc f1-l fw1 mt5-l">The page you seek does not exist.</h2>
         </header>
-        <p class="fw1 i tc mt4 mt5-l f4 f3-l">Are you looking for one of these?</p>
+        <p class="fw3 i tc mt4 mt5-l f4 f3-l">Are you looking for one of these?</p>
         <ul class="list tc pl0 w-100 mt5">
           <li class="dib"><a class="f5 f4-ns link color-green db pv2 ph3 hover-color-green-hover" href="/">Dat Home</a></li>
           <li class="dib"><a class="f5 f4-ns link color-blue db pv2 ph3 hover-color-blue-hover" href="/about">About</a></li>
