+++ package.json
@@ -18,11 +18,6 @@
     "fa-dev": "fair-analytics --memory",
     "test": "standard && dependency-check ./package.json"
   },
-  "sheetify": {
-    "transform": [
-      "sheetify-nested"
-    ]
-  },
   "devDependencies": {
     "bankai": "^9.14.0",
     "dependency-check": "^2.9.1",
@@ -31,12 +26,11 @@
   },
   "dependencies": {
     "choo": "^6.0.0",
-    "choo-log": "^8.0.0",
+    "choo-devtools": "^2.5.1",
     "dat-colors": "^3.5.1",
     "dat-icons": "^2.5.0",
     "microanalytics": "^1.0.1",
     "sheetify": "^7.3.2",
-    "sheetify-nested": "^1.0.2",
     "tachyons": "^4.10.0"
   }
 }
